[IMP] mail,*: remove useless field `is_master_data`
This field is a non-sense. By using the same logic, it should be added
to any low level model.
Replace the check by a simple verification of presence of an XMLID.

A more generic opt-in solution should be integrated into ORM.

Partially revert commits 0db0e66e96a8f6b393565da6fbb8dce25ec960b7 and
bbd64c22abc19558c1d83a892fc426bf99a094ce.

See #29257 odoo/enterprise#3550

+++ addons/mail/models/mail_activity.py
@@ -78,7 +78,6 @@ class MailActivityType(models.Model):
     initial_res_model_id = fields.Many2one('ir.model', 'Initial model', compute="_compute_initial_res_model_id", store=False,
             help='Technical field to keep trace of the model at the beginning of the edition for UX related behaviour')
     res_model_change = fields.Boolean(string="Model has change", help="Technical field for UX related behaviour", default=False, store=False)
-    is_master_data = fields.Boolean(string="Master data", help="This field is used to prevent master data from the deletion.")
 
     @api.onchange('res_model_id')
     def _onchange_res_model_id(self):
@@ -91,9 +90,8 @@ class MailActivityType(models.Model):
 
     @api.multi
     def unlink(self):
-        for activity_type in self:
-            if activity_type.is_master_data:
-                raise exceptions.ValidationError("You can not delete activity type that are used as master data.")
+        if any(self.get_external_id().values()):
+            raise exceptions.ValidationError("You can not delete activity type that are used as master data.")
         return super(MailActivityType, self).unlink()
